:truck: Move state store unit tests to folder
Fix typos and its refernce
